+++ R/generate_pattern.R
@@ -6,7 +6,7 @@
 #' \code{Theta} input with more than 1 row will generate a matrix of responses for
 #' running independent CAT session when passed to \code{mirtCAT(..., local_pattern)}.
 #' 
-#' @param mirt_object single group object defined by the \code{mirt} package
+#' @param mo single group object defined by the \code{mirt} package
 #'
 #' @param Theta a numeric vector indicating the latent theta values for a single person
 #' 
@@ -22,33 +22,33 @@
 #' 
 #' # return real response vector given choices and (optional) answers 
 #' pat <- generate_pattern(mod, Theta = 0, df=df)
-#' # mirtCAT(df, mirt_object=mod, local_pattern = pat)
+#' # mirtCAT(df, mo=mod, local_pattern = pat)
 #' 
 #' # generate single pattern observed in dataset used to define mod
 #' pat2 <- generate_pattern(mod, Theta = 0)
-#' # mirtCAT(mirt_object=mod, local_pattern = pat2)
+#' # mirtCAT(mo=mod, local_pattern = pat2)
 #'
 #' # generate multiple patterns to be analyzed independently 
 #' pat3 <- generate_pattern(mod, Theta = matrix(c(0, 2, -2), 3))
-#' # mirtCAT(mirt_object=mod, local_pattern = pat3)
+#' # mirtCAT(mo=mod, local_pattern = pat3)
 #' 
 #' }
-generate_pattern <- function(mirt_object, Theta, df = NULL){
+generate_pattern <- function(mo, Theta, df = NULL){
     fn <- function(p, ns) sample(1L:ns, 1L, prob = p) - 1
-    nitems <- ncol(mirt_object@Data$data)
+    nitems <- ncol(mo@Data$data)
     if(!is.matrix(Theta)) Theta <- matrix(Theta, 1L)
     if(nrow(Theta) > 1L && !is.null(df))
         stop('df argument only used for generating single-case response patterns')
     N <- nrow(Theta)
     pattern <- matrix(0L, N, nitems)
     if(is.null(df)){
-        K <- mirt_object@Data$K
+        K <- mo@Data$K
         for(i in 1L:nitems){
-            ii <- extract.item(mirt_object, i)
+            ii <- extract.item(mo, i)
             P <- probtrace(ii, Theta)
             pattern[,i] <- apply(P, 1L, fn, ns = ncol(P))
         }
-        return(t(t(pattern) + mirt_object@Data$mins))
+        return(t(t(pattern) + mo@Data$mins))
     } else {
         choices <- df[,grepl('Option', colnames(df))]
         item_answers <- df[,grepl('Answer', colnames(df))]
@@ -58,7 +58,7 @@ generate_pattern <- function(mirt_object, Theta, df = NULL){
     ret <- character(nitems)
     has_item_answers <- length(item_answers) > 0L
     for(i in 1L:nitems){
-        ii <- extract.item(mirt_object, i)
+        ii <- extract.item(mo, i)
         P <- probtrace(ii, Theta)
         uniq <- 1L:ncol(P) - 1L
         pattern[i] <- sample(uniq, 1L, prob = P)